+++ src/TinyGsmClientSaraR4.h
@@ -483,7 +483,7 @@ class TinyGsmSaraR4 : public TinyGsmModem<TinyGsmSaraR4>,
   }
 
   inline bool getUbloxLocation(int8_t sensor, float* lat, float* lon,
-                               float* speed = 0, int* alt = 0, int* vsat = 0,
+                               float* speed = 0, float* alt = 0, int* vsat = 0,
                                int* usat = 0, float* accuracy = 0,
                                int* year = 0, int* month = 0, int* day = 0,
                                int* hour = 0, int* minute = 0,
@@ -556,7 +556,7 @@ class TinyGsmSaraR4 : public TinyGsmModem<TinyGsmSaraR4>,
     if (lat != NULL) *lat = ilat;
     if (lon != NULL) *lon = ilon;
     if (speed != NULL) *speed = ispeed;
-    if (alt != NULL) *alt = static_cast<int>(ialt);
+    if (alt != NULL) *alt = ialt;
     if (vsat != NULL) *vsat = 0;  // Number of satellites viewed not reported;
     if (usat != NULL) *usat = iusat;
     if (accuracy != NULL) *accuracy = iaccuracy;
@@ -578,7 +578,7 @@ class TinyGsmSaraR4 : public TinyGsmModem<TinyGsmSaraR4>,
     return getUbloxLocation(2, lat, lon, 0, 0, 0, 0, accuracy, year, month, day,
                             hour, minute, second);
   }
-  bool getGPSImpl(float* lat, float* lon, float* speed = 0, int* alt = 0,
+  bool getGPSImpl(float* lat, float* lon, float* speed = 0, float* alt = 0,
                   int* vsat = 0, int* usat = 0, float* accuracy = 0,
                   int* year = 0, int* month = 0, int* day = 0, int* hour = 0,
                   int* minute = 0, int* second = 0) {
